# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your Application API document",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden,Try again!"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": "403 "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "register"
                ],
                "summary": "Register",
                "description": "Register by email, password",
                "operationId": "register",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "A@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123abc"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Nguyen Van A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Duplicate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User1 is duplicate.Try Agian!"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": "201"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/list": {
            "get": {
                "tags": [
                    "list"
                ],
                "summary": "List user",
                "description": "Get list User",
                "operationId": "list",
                "responses": {
                    "404": {
                        "description": "Fail to filter token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "null"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "filter token fail"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/changePassword": {
            "post": {
                "tags": [
                    "changePassword"
                ],
                "summary": "Change",
                "description": "Register by email, password",
                "operationId": "changePassword",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "A@mail.com"
                                    },
                                    "old_pass": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123abc"
                                    },
                                    "new_pass": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "aokok123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "wrong infomations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "eamil or password is not correctly.Try Agian!"
                                        },
                                        "status": {
                                            "type": "int",
                                            "example": "201"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}